fix mock stop, rename post-processing, fix eslint
