+++ lib/postprocessing/image.js
@@ -12,8 +12,7 @@
 
 'use strict';
 
-const IMAGEMAGICK = true;
-const gm = require("gm").subClass({ imageMagick: IMAGEMAGICK });
+const gm = require("gm").subClass({ imageMagick: true });
 const { execSync } = require("child_process");
 const fs = require("fs-extra");
 const path = require("path");
@@ -146,7 +145,7 @@ function applyImageOperations(img, instructions, operations){
 }
 
 async function logImageMagickVersion() {
-    const version = IMAGEMAGICK ? execSync("identify --version") : execSync("gm version");
+    const version = execSync("identify --version");
     console.log(version.toString());
 }
