+++ lib/worker.js
@@ -20,7 +20,7 @@ const { createDirectories, cleanupDirectories } = require('./prepare');
 const { getSource, putRendition } = require('./storage');
 const Rendition = require('./rendition');
 
-const { imageProcess } = require('./post-processing/image');
+const { imageProcess } = require('./postprocessing/image');
 const { validateParameters } = require('./validate');
 const Timer = require('./utils/timer');
 const { Sampler } = require('@adobe/metrics-sampler');
@@ -78,10 +78,6 @@ class AssetComputeWorker {
         if (!process.env.DISABLE_IO_EVENTS_ON_TIMEOUT) {
             this.actionTimeoutId = this.sendEventsBeforeActionTimeout();
         }
-
-        // if (options.postProcessingEngine) {
-        //     this.postProcessingEngine = options.postProcessingEngine;
-        // }
 
         this.renditionErrors = [];
 
@@ -317,7 +313,6 @@ class AssetComputeWorker {
 
         console.log(`Post processing rendition ${rendition.name} => ${postProcessedRendition.name}`);
 
-        // const postProcessingEngine = this.postProcessingEngine || imageProcess;
         await imageProcess(rendition.path, postProcessedRendition.path, rendition.instructions);
 
         // point to proper rendition once postprocessing done
