+++ test-worker/package.json
@@ -5,9 +5,6 @@
   "scripts": {
     "test": "adobe-asset-compute test-worker"
   },
-  "dependencies": {
-    "@adobe/asset-compute-sdk": "https://github.com:adobe/asset-compute-sdk.git#mob/NUI-668"
-  },
   "devDependencies": {
     "@adobe/aio-cli-plugin-asset-compute": "^1.3.8"
   }
