+++ test/rendition.postprocess.test.js
@@ -29,7 +29,7 @@ const BASE64_RENDITION_JPG = 'ZmZkOGZmZTAwMDEwNGE0NjQ5NDYwMDAxMDEwMDAwMDEwMDAxMD
 const BASE64_RENDITION_PNG = "ODk1MDRlNDcwZDBhMWEwYTAwMDAwMDBkNDk0ODQ0NTIwMDAwMDAwYTAwMDAwMDA2MDgwNjAwMDAwMGZhZjAwZmM2MDAwMDAwMDE3MzUyNDc0MjAwYWVjZTFjZTkwMDAwMDA1MDY1NTg0OTY2NGQ0ZDAwMmEwMDAwMDAwODAwMDIwMTEyMDAwMzAwMDAwMDAxMDAwMTAwMDA4NzY5MDAwNDAwMDAwMDAxMDAwMDAwMjYwMDAwMDAwMDAwMDNhMDAxMDAwMzAwMDAwMDAxMDAwMTAwMDBhMDAyMDAwNDAwMDAwMDAxMDAwMDAwMGFhMDAzMDAwNDAwMDAwMDAxMDAwMDAwMDYwMDAwMDAwMGVjNTU3YmVhMDAwMDAxNTk2OTU0NTg3NDU4NGQ0YzNhNjM2ZjZkMmU2MTY0NmY2MjY1MmU3ODZkNzAwMDAwMDAwMDAwM2M3ODNhNzg2ZDcwNmQ2NTc0NjEyMDc4NmQ2YzZlNzMzYTc4M2QyMjYxNjQ2ZjYyNjUzYTZlNzMzYTZkNjU3NDYxMmYyMjIwNzgzYTc4NmQ3MDc0NmIzZDIyNTg0ZDUwMjA0MzZmNzI2NTIwMzUyZTM0MmUzMDIyM2UwYTIwMjAyMDNjNzI2NDY2M2E1MjQ0NDYyMDc4NmQ2YzZlNzMzYTcyNjQ2NjNkMjI2ODc0NzQ3MDNhMmYyZjc3Nzc3NzJlNzczMzJlNmY3MjY3MmYzMTM5MzkzOTJmMzAzMjJmMzIzMjJkNzI2NDY2MmQ3Mzc5NmU3NDYxNzgyZDZlNzMyMzIyM2UwYTIwMjAyMDIwMjAyMDNjNzI2NDY2M2E0NDY1NzM2MzcyNjk3MDc0Njk2ZjZlMjA3MjY0NjYzYTYxNjI2Zjc1NzQzZDIyMjIwYTIwMjAyMDIwMjAyMDIwMjAyMDIwMjAyMDc4NmQ2YzZlNzMzYTc0Njk2NjY2M2QyMjY4NzQ3NDcwM2EyZjJmNmU3MzJlNjE2NDZmNjI2NTJlNjM2ZjZkMmY3NDY5NjY2NjJmMzEyZTMwMmYyMjNlMGEyMDIwMjAyMDIwMjAyMDIwMjAzYzc0Njk2NjY2M2E0ZjcyNjk2NTZlNzQ2MTc0Njk2ZjZlM2UzMTNjMmY3NDY5NjY2NjNhNGY3MjY5NjU2ZTc0NjE3NDY5NmY2ZTNlMGEyMDIwMjAyMDIwMjAzYzJmNzI2NDY2M2E0NDY1NzM2MzcyNjk3MDc0Njk2ZjZlM2UwYTIwMjAyMDNjMmY3MjY0NjYzYTUyNDQ0NjNlMGEzYzJmNzgzYTc4NmQ3MDZkNjU3NDYxM2UwYTRjYzIyNzU5MDAwMDAwYmY0OTQ0NDE1NDA4MWQzZDhlY2I2YWMyNDAxNDg2YmYzMTMzNTQ4NTc4MmQyMWMxZDQ1MjE0N2M4MGFlZjU5NTdkMGYxZjQyOTAyMmQ5YzUzNDVhOWE3NGU4OGM2NzIyNzgxNjY3ZjFkZjk1NGEzZTNkNzI0YWJlYjMxNmRhM2ZmODc3MTA0NWY0YjRjNmY3MGNkZTdiYjQ2ZjFhNTEyOTdjMjAwNTQwYzhmZTY4ZDgxOWUzMzhlNmI3YjEwZjYxN2YzNjE1NzNkNGE1YmNjZWU2MmNiMjhjYjc0NTRlOWVlNzRjYzcxMzU2ZWI0ZDI4NDQ2Zjc3M2I4MjMzNGI1MzNlOTZlZjlkMjg0ZDEyNmVmNTU1OGExNDkzZjExYzYzMGM2YTdmMzhmYjUwNWI5NjI1NDY2YTVmYjQ4MGIyYzBiNjJkNTYzNjlmOGU1ZjM4ZTdkMGY1Y2Y4ZDZiZjU0ZDUxMTQ1NGU1ODVlMTYwZjAxNDg3YzRiYWFhYmI4ZDc3Y2I1MjRkOTA4NmI3Y2IyNjAwMDAwMDAwNDk0NTRlNDRhZTQyNjA4Mg==";
 const BASE64_RENDITION_TIFF = "NGQ0ZDAwMmEwMDAwMDBmODAxMTYzM2ZmMDExNjMzZmYwMjE1MzNmZjAwMTAyZWZmMDAwOTI4ZmYwMDA2MjZmZjAwMDkyOWZmMDIwZDJkZmYwMTBmMzJmZjAwMGUzMWZmMDAwZTJiZmYwMDBmMmNmZjAwMGMyYWZmMDAwYjI5ZmYwMDBkMmNmZjA4MTczNmZmMDkxNjM2ZmYxNjIzNDNmZjBkMWIzZWZmMGMxYTNkZmYwODIwM2NmZjBiMjMzZmZmMGIyMDNkZmYxZjMyNTBmZjNhNGI2OWZmNTc2Njg1ZmY3MzgyYTFmZjg0OTFiMWZmYTdiNWQ4ZmZhN2I1ZDhmZjQ1NWU3YWZmNTI2Yjg3ZmY2YjgzOWZmZjhjYTFiZWZmYTliY2RhZmZjMWQyZjBmZmI1YzRlM2ZmYjZjNWU0ZmZjNmQ0ZjdmZmNiZDlmY2ZmYWZjYWU1ZmZhZWM3ZTNmZjljYjVkMWZmYTFiOWQ1ZmZhOGJkZGFmZmE5YmNkYWZmYTRiNWQzZmY5ZmIwY2VmZjgwOGViMWZmN2U4Y2FmZmZhMWJmZDlmZjk0YWZjYWZmN2I5NmIxZmY2Yjg0YTBmZjc1OGRhOWZmN2M5MWFlZmY3ZDkwYWVmZjdlOTFhZmZmNzA3ZWExZmY2ZjdkYTBmZjAwMTAwMTAwMDAwMzAwMDAwMDAxMDAwYTAwMDAwMTAxMDAwMzAwMDAwMDAxMDAwNjAwMDAwMTAyMDAwMzAwMDAwMDA0MDAwMDAxYmUwMTAzMDAwMzAwMDAwMDAxMDAwMTAwMDAwMTA2MDAwMzAwMDAwMDAxMDAwMjAwMDAwMTBhMDAwMzAwMDAwMDAxMDAwMTAwMDAwMTExMDAwNDAwMDAwMDAxMDAwMDAwMDgwMTEyMDAwMzAwMDAwMDAxMDAwMTAwMDAwMTE1MDAwMzAwMDAwMDAxMDAwNDAwMDAwMTE2MDAwMzAwMDAwMDAxMDAwNjAwMDAwMTE3MDAwNDAwMDAwMDAxMDAwMDAwZjAwMTFjMDAwMzAwMDAwMDAxMDAwMTAwMDAwMTI4MDAwMzAwMDAwMDAxMDAwMjAwMDAwMTUyMDAwMzAwMDAwMDAxMDAwMTAwMDAwMTUzMDAwMzAwMDAwMDA0MDAwMDAxYzY4NzczMDAwNzAwMDAwYzQ4MDAwMDAxY2UwMDAwMDAwMDAwMDgwMDA4MDAwODAwMDgwMDAxMDAwMTAwMDEwMDAxMDAwMDBjNDg0YzY5NmU2ZjAyMTAwMDAwNmQ2ZTc0NzI1MjQ3NDIyMDU4NTk1YTIwMDdjZTAwMDIwMDA5MDAwNjAwMzEwMDAwNjE2MzczNzA0ZDUzNDY1NDAwMDAwMDAwNDk0NTQzMjA3MzUyNDc0MjAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDBmNmQ2MDAwMTAwMDAwMDAwZDMyZDQ4NTAyMDIwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDExNjM3MDcyNzQwMDAwMDE1MDAwMDAwMDMzNjQ2NTczNjMwMDAwMDE4NDAwMDAwMDZjNzc3NDcwNzQwMDAwMDFmMDAwMDAwMDE0NjI2YjcwNzQwMDAwMDIwNDAwMDAwMDE0NzI1ODU5NWEwMDAwMDIxODAwMDAwMDE0Njc1ODU5NWEwMDAwMDIyYzAwMDAwMDE0NjI1ODU5NWEwMDAwMDI0MDAwMDAwMDE0NjQ2ZDZlNjQwMDAwMDI1NDAwMDAwMDcwNjQ2ZDY0NjQwMDAwMDJjNDAwMDAwMDg4NzY3NTY1NjQwMDAwMDM0YzAwMDAwMDg2NzY2OTY1NzcwMDAwMDNkNDAwMDAwMDI0NmM3NTZkNjkwMDAwMDNmODAwMDAwMDE0NmQ2NTYxNzMwMDAwMDQwYzAwMDAwMDI0NzQ2NTYzNjgwMDAwMDQzMDAwMDAwMDBjNzI1NDUyNDMwMDAwMDQzYzAwMDAwODBjNjc1NDUyNDMwMDAwMDQzYzAwMDAwODBjNjI1NDUyNDMwMDAwMDQzYzAwMDAwODBjNzQ2NTc4NzQwMDAwMDAwMDQzNmY3MDc5NzI2OTY3Njg3NDIwMjg2MzI5MjAzMTM5MzkzODIwNDg2NTc3NmM2NTc0NzQyZDUwNjE2MzZiNjE3MjY0MjA0MzZmNmQ3MDYxNmU3OTAwMDA2NDY1NzM2MzAwMDAwMDAwMDAwMDAwMTI3MzUyNDc0MjIwNDk0NTQzMzYzMTM5MzYzNjJkMzIyZTMxMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDEyNzM1MjQ3NDIyMDQ5NDU0MzM2MzEzOTM2MzYyZDMyMmUzMTAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDA1ODU5NWEyMDAwMDAwMDAwMDAwMGYzNTEwMDAxMDAwMDAwMDExNmNjNTg1OTVhMjAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDU4NTk1YTIwMDAwMDAwMDAwMDAwNmZhMjAwMDAzOGY1MDAwMDAzOTA1ODU5NWEyMDAwMDAwMDAwMDAwMDYyOTkwMDAwYjc4NTAwMDAxOGRhNTg1OTVhMjAwMDAwMDAwMDAwMDAyNGEwMDAwMDBmODQwMDAwYjZjZjY0NjU3MzYzMDAwMDAwMDAwMDAwMDAxNjQ5NDU0MzIwNjg3NDc0NzAzYTJmMmY3Nzc3NzcyZTY5NjU2MzJlNjM2ODAwMDAwMDAwMDAwMDAwMDAwMDAwMDAxNjQ5NDU0MzIwNjg3NDc0NzAzYTJmMmY3Nzc3NzcyZTY5NjU2MzJlNjM2ODAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwNjQ2NTczNjMwMDAwMDAwMDAwMDAwMDJlNDk0NTQzMjAzNjMxMzkzNjM2MmQzMjJlMzEyMDQ0NjU2NjYxNzU2Yzc0MjA1MjQ3NDIyMDYzNmY2YzZmNzU3MjIwNzM3MDYxNjM2NTIwMmQyMDczNTI0NzQyMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDJlNDk0NTQzMjAzNjMxMzkzNjM2MmQzMjJlMzEyMDQ0NjU2NjYxNzU2Yzc0MjA1MjQ3NDIyMDYzNmY2YzZmNzU3MjIwNzM3MDYxNjM2NTIwMmQyMDczNTI0NzQyMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDA2NDY1NzM2MzAwMDAwMDAwMDAwMDAwMmM1MjY1NjY2NTcyNjU2ZTYzNjUyMDU2Njk2NTc3Njk2ZTY3MjA0MzZmNmU2NDY5NzQ2OTZmNmUyMDY5NmUyMDQ5NDU0MzM2MzEzOTM2MzYyZDMyMmUzMTAwMDAwMDAwMDAwMDAwMDAwMDAwMDAyYzUyNjU2NjY1NzI2NTZlNjM2NTIwNTY2OTY1Nzc2OTZlNjcyMDQzNmY2ZTY0Njk3NDY5NmY2ZTIwNjk2ZTIwNDk0NTQzMzYzMTM5MzYzNjJkMzIyZTMxMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDc2Njk2NTc3MDAwMDAwMDAwMDEzYTRmZTAwMTQ1ZjJlMDAxMGNmMTQwMDAzZWRjYzAwMDQxMzBiMDAwMzVjOWUwMDAwMDAwMTU4NTk1YTIwMDAwMDAwMDAwMDRjMDk1NjAwNTAwMDAwMDA1NzFmZTc2ZDY1NjE3MzAwMDAwMDAwMDAwMDAwMDEwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMDAwMjhmMDAwMDAwMDI3MzY5NjcyMDAwMDAwMDAwNDM1MjU0MjA2Mzc1NzI3NjAwMDAwMDAwMDAwMDA0MDAwMDAwMDAwNTAwMGEwMDBmMDAxNDAwMTkwMDFlMDAyMzAwMjgwMDJkMDAzMjAwMzcwMDNiMDA0MDAwNDUwMDRhMDA0ZjAwNTQwMDU5MDA1ZTAwNjMwMDY4MDA2ZDAwNzIwMDc3MDA3YzAwODEwMDg2MDA4YjAwOTAwMDk1MDA5YTAwOWYwMGE0MDBhOTAwYWUwMGIyMDBiNzAwYmMwMGMxMDBjNjAwY2IwMGQwMDBkNTAwZGIwMGUwMDBlNTAwZWIwMGYwMDBmNjAwZmIwMTAxMDEwNzAxMGQwMTEzMDExOTAxMWYwMTI1MDEyYjAxMzIwMTM4MDEzZTAxNDUwMTRjMDE1MjAxNTkwMTYwMDE2NzAxNmUwMTc1MDE3YzAxODMwMThiMDE5MjAxOWEwMWExMDFhOTAxYjEwMWI5MDFjMTAxYzkwMWQxMDFkOTAxZTEwMWU5MDFmMjAxZmEwMjAzMDIwYzAyMTQwMjFkMDIyNjAyMmYwMjM4MDI0MTAyNGIwMjU0MDI1ZDAyNjcwMjcxMDI3YTAyODQwMjhlMDI5ODAyYTIwMmFjMDJiNjAyYzEwMmNiMDJkNTAyZTAwMmViMDJmNTAzMDAwMzBiMDMxNjAzMjEwMzJkMDMzODAzNDMwMzRmMDM1YTAzNjYwMzcyMDM3ZTAzOGEwMzk2MDNhMjAzYWUwM2JhMDNjNzAzZDMwM2UwMDNlYzAzZjkwNDA2MDQxMzA0MjAwNDJkMDQzYjA0NDgwNDU1MDQ2MzA0NzEwNDdlMDQ4YzA0OWEwNGE4MDRiNjA0YzQwNGQzMDRlMTA0ZjAwNGZlMDUwZDA1MWMwNTJiMDUzYTA1NDkwNTU4MDU2NzA1NzcwNTg2MDU5NjA1YTYwNWI1MDVjNTA1ZDUwNWU1MDVmNjA2MDYwNjE2MDYyNzA2MzcwNjQ4MDY1OTA2NmEwNjdiMDY4YzA2OWQwNmFmMDZjMDA2ZDEwNmUzMDZmNTA3MDcwNzE5MDcyYjA3M2QwNzRmMDc2MTA3NzQwNzg2MDc5OTA3YWMwN2JmMDdkMjA3ZTUwN2Y4MDgwYjA4MWYwODMyMDg0NjA4NWEwODZlMDg4MjA4OTYwOGFhMDhiZTA4ZDIwOGU3MDhmYjA5MTAwOTI1MDkzYTA5NGYwOTY0MDk3OTA5OGYwOWE0MDliYTA5Y2YwOWU1MDlmYjBhMTEwYTI3MGEzZDBhNTQwYTZhMGE4MTBhOTgwYWFlMGFjNTBhZGMwYWYzMGIwYjBiMjIwYjM5MGI1MTBiNjkwYjgwMGI5ODBiYjAwYmM4MGJlMTBiZjkwYzEyMGMyYTBjNDMwYzVjMGM3NTBjOGUwY2E3MGNjMDBjZDkwY2YzMGQwZDBkMjYwZDQwMGQ1YTBkNzQwZDhlMGRhOTBkYzMwZGRlMGRmODBlMTMwZTJlMGU0OTBlNjQwZTdmMGU5YjBlYjYwZWQyMGVlZTBmMDkwZjI1MGY0MTBmNWUwZjdhMGY5NjBmYjMwZmNmMGZlYzEwMDkxMDI2MTA0MzEwNjExMDdlMTA5YjEwYjkxMGQ3MTBmNTExMTMxMTMxMTE0ZjExNmQxMThjMTFhYTExYzkxMWU4MTIwNzEyMjYxMjQ1MTI2NDEyODQxMmEzMTJjMzEyZTMxMzAzMTMyMzEzNDMxMzYzMTM4MzEzYTQxM2M1MTNlNTE0MDYxNDI3MTQ0OTE0NmExNDhiMTRhZDE0Y2UxNGYwMTUxMjE1MzQxNTU2MTU3ODE1OWIxNWJkMTVlMDE2MDMxNjI2MTY0OTE2NmMxNjhmMTZiMjE2ZDYxNmZhMTcxZDE3NDExNzY1MTc4OTE3YWUxN2QyMTdmNzE4MWIxODQwMTg2NTE4OGExOGFmMThkNTE4ZmExOTIwMTk0NTE5NmIxOTkxMTliNzE5ZGQxYTA0MWEyYTFhNTExYTc3MWE5ZTFhYzUxYWVjMWIxNDFiM2IxYjYzMWI4YTFiYjIxYmRhMWMwMjFjMmExYzUyMWM3YjFjYTMxY2NjMWNmNTFkMWUxZDQ3MWQ3MDFkOTkxZGMzMWRlYzFlMTYxZTQwMWU2YTFlOTQxZWJlMWVlOTFmMTMxZjNlMWY2OTFmOTQxZmJmMWZlYTIwMTUyMDQxMjA2YzIwOTgyMGM0MjBmMDIxMWMyMTQ4MjE3NTIxYTEyMWNlMjFmYjIyMjcyMjU1MjI4MjIyYWYyMmRkMjMwYTIzMzgyMzY2MjM5NDIzYzIyM2YwMjQxZjI0NGQyNDdjMjRhYjI0ZGEyNTA5MjUzODI1NjgyNTk3MjVjNzI1ZjcyNjI3MjY1NzI2ODcyNmI3MjZlODI3MTgyNzQ5Mjc3YTI3YWIyN2RjMjgwZDI4M2YyODcxMjhhMjI4ZDQyOTA2MjkzODI5NmIyOTlkMjlkMDJhMDIyYTM1MmE2ODJhOWIyYWNmMmIwMjJiMzYyYjY5MmI5ZDJiZDEyYzA1MmMzOTJjNmUyY2EyMmNkNzJkMGMyZDQxMmQ3NjJkYWIyZGUxMmUxNjJlNGMyZTgyMmViNzJlZWUyZjI0MmY1YTJmOTEyZmM3MmZmZTMwMzUzMDZjMzBhNDMwZGIzMTEyMzE0YTMxODIzMWJhMzFmMjMyMmEzMjYzMzI5YjMyZDQzMzBkMzM0NjMzN2YzM2I4MzNmMTM0MmIzNDY1MzQ5ZTM0ZDgzNTEzMzU0ZDM1ODczNWMyMzVmZDM2MzczNjcyMzZhZTM2ZTkzNzI0Mzc2MDM3OWMzN2Q3MzgxNDM4NTAzODhjMzhjODM5MDUzOTQyMzk3ZjM5YmMzOWY5M2EzNjNhNzQzYWIyM2FlZjNiMmQzYjZiM2JhYTNiZTgzYzI3M2M2NTNjYTQzY2UzM2QyMjNkNjEzZGExM2RlMDNlMjAzZTYwM2VhMDNlZTAzZjIxM2Y2MTNmYTIzZmUyNDAyMzQwNjQ0MGE2NDBlNzQxMjk0MTZhNDFhYzQxZWU0MjMwNDI3MjQyYjU0MmY3NDMzYTQzN2Q0M2MwNDQwMzQ0NDc0NDhhNDRjZTQ1MTI0NTU1NDU5YTQ1ZGU0NjIyNDY2NzQ2YWI0NmYwNDczNTQ3N2I0N2MwNDgwNTQ4NGI0ODkxNDhkNzQ5MWQ0OTYzNDlhOTQ5ZjA0YTM3NGE3ZDRhYzQ0YjBjNGI1MzRiOWE0YmUyNGMyYTRjNzI0Y2JhNGQwMjRkNGE0ZDkzNGRkYzRlMjU0ZTZlNGViNzRmMDA0ZjQ5NGY5MzRmZGQ1MDI3NTA3MTUwYmI1MTA2NTE1MDUxOWI1MWU2NTIzMTUyN2M1MmM3NTMxMzUzNWY1M2FhNTNmNjU0NDI1NDhmNTRkYjU1Mjg1NTc1NTVjMjU2MGY1NjVjNTZhOTU2Zjc1NzQ0NTc5MjU3ZTA1ODJmNTg3ZDU4Y2I1OTFhNTk2OTU5Yjg1YTA3NWE1NjVhYTY1YWY1NWI0NTViOTU1YmU1NWMzNTVjODY1Y2Q2NWQyNzVkNzg1ZGM5NWUxYTVlNmM1ZWJkNWYwZjVmNjE1ZmIzNjAwNTYwNTc2MGFhNjBmYzYxNGY2MWEyNjFmNTYyNDk2MjljNjJmMDYzNDM2Mzk3NjNlYjY0NDA2NDk0NjRlOTY1M2Q2NTkyNjVlNzY2M2Q2NjkyNjZlODY3M2Q2NzkzNjdlOTY4M2Y2ODk2NjhlYzY5NDM2OTlhNjlmMTZhNDg2YTlmNmFmNzZiNGY2YmE3NmJmZjZjNTc2Y2FmNmQwODZkNjA2ZGI5NmUxMjZlNmI2ZWM0NmYxZTZmNzg2ZmQxNzAyYjcwODY3MGUwNzEzYTcxOTU3MWYwNzI0YjcyYTY3MzAxNzM1ZDczYjg3NDE0NzQ3MDc0Y2M3NTI4NzU4NTc1ZTE3NjNlNzY5Yjc2Zjg3NzU2NzdiMzc4MTE3ODZlNzhjYzc5MmE3OTg5NzllNzdhNDY3YWE1N2IwNDdiNjM3YmMyN2MyMTdjODE3Y2UxN2Q0MTdkYTE3ZTAxN2U2MjdlYzI3ZjIzN2Y4NDdmZTU4MDQ3ODBhODgxMGE4MTZiODFjZDgyMzA4MjkyODJmNDgzNTc4M2JhODQxZDg0ODA4NGUzODU0Nzg1YWI4NjBlODY3Mjg2ZDc4NzNiODc5Zjg4MDQ4ODY5ODhjZTg5MzM4OTk5ODlmZThhNjQ4YWNhOGIzMDhiOTY4YmZjOGM2MzhjY2E4ZDMxOGQ5ODhkZmY4ZTY2OGVjZThmMzY4ZjllOTAwNjkwNmU5MGQ2OTEzZjkxYTg5MjExOTI3YTkyZTM5MzRkOTNiNjk0MjA5NDhhOTRmNDk1NWY5NWM5OTYzNDk2OWY5NzBhOTc3NTk3ZTA5ODRjOThiODk5MjQ5OTkwOTlmYzlhNjg5YWQ1OWI0MjliYWY5YzFjOWM4OTljZjc5ZDY0OWRkMjllNDA5ZWFlOWYxZDlmOGI5ZmZhYTA2OWEwZDhhMTQ3YTFiNmEyMjZhMjk2YTMwNmEzNzZhM2U2YTQ1NmE0YzdhNTM4YTVhOWE2MWFhNjhiYTZmZGE3NmVhN2UwYTg1MmE4YzRhOTM3YTlhOWFhMWNhYThmYWIwMmFiNzVhYmU5YWM1Y2FjZDBhZDQ0YWRiOGFlMmRhZWExYWYxNmFmOGJiMDAwYjA3NWIwZWFiMTYwYjFkNmIyNGJiMmMyYjMzOGIzYWViNDI1YjQ5Y2I1MTNiNThhYjYwMWI2NzliNmYwYjc2OGI3ZTBiODU5YjhkMWI5NGFiOWMyYmEzYmJhYjViYjJlYmJhN2JjMjFiYzliYmQxNWJkOGZiZTBhYmU4NGJlZmZiZjdhYmZmNWMwNzBjMGVjYzE2N2MxZTNjMjVmYzJkYmMzNThjM2Q0YzQ1MWM0Y2VjNTRiYzVjOGM2NDZjNmMzYzc0MWM3YmZjODNkYzhiY2M5M2FjOWI5Y2EzOGNhYjdjYjM2Y2JiNmNjMzVjY2I1Y2QzNWNkYjVjZTM2Y2ViNmNmMzdjZmI4ZDAzOWQwYmFkMTNjZDFiZWQyM2ZkMmMxZDM0NGQzYzZkNDQ5ZDRjYmQ1NGVkNWQxZDY1NWQ2ZDhkNzVjZDdlMGQ4NjRkOGU4ZDk2Y2Q5ZjFkYTc2ZGFmYmRiODBkYzA1ZGM4YWRkMTBkZDk2ZGUxY2RlYTJkZjI5ZGZhZmUwMzZlMGJkZTE0NGUxY2NlMjUzZTJkYmUzNjNlM2ViZTQ3M2U0ZmNlNTg0ZTYwZGU2OTZlNzFmZTdhOWU4MzJlOGJjZTk0NmU5ZDBlYTViZWFlNWViNzBlYmZiZWM4NmVkMTFlZDljZWUyOGVlYjRlZjQwZWZjY2YwNThmMGU1ZjE3MmYxZmZmMjhjZjMxOWYzYTdmNDM0ZjRjMmY1NTBmNWRlZjY2ZGY2ZmJmNzhhZjgxOWY4YThmOTM4ZjljN2ZhNTdmYWU3ZmI3N2ZjMDdmYzk4ZmQyOWZkYmFmZTRiZmVkY2ZmNmRmZmZm";
 
-mockRequire("../lib/post-processing/image", {
+mockRequire("../lib/postprocessing/image", {
     imageProcess: async function(infile, outfile, instructions) {
         console.log('mocked image post processing', outfile, infile);
         if (instructions.shouldFail) {
@@ -68,7 +68,7 @@ describe("imagePostProcess", () => {
     });
 
     after(() => {
-        mockRequire.stop('./lib/post-processing/image');
+        mockRequire.stop('../lib/postprocessing/image');
     });
 
     it('should convert PNG to JPG - end to end test', async () => {
